Refactored Helper Method into helper class
