+++ scripts/examples/map/performanceTest/generateTestData.py
@@ -432,13 +432,13 @@ if __name__ == "__main__":
     }
 
     matrix_sizes_500mb_1gb = {   
-    'f32': [8192, 16384],
-    'f64': [5792, 11585],
-    'int32': [8192, 16384],
-    'int64': [5792, 11585],
-    'int8': [16384, 32768],
-    'uint64': [5792, 11585],
-    'uint8': [16384, 32768]
+    'f32': [100,1000,8192, 16384],
+    'f64': [100,1000,5792, 11585],
+    'int32': [100,1000,8192, 16384],
+    'int64': [100,1000,5792, 11585],
+    'int8': [100,1000,16384, 32768],
+    'uint64': [100,1000,5792, 11585],
+    'uint8': [100,1000,16384, 32768]
     }
 
     matrix_sizes_250mb_500mb = {
