+++ src/runtime/local/kernels/MAP_EXTERNAL/CtypesMapKernel_sharedMem_address.h
@@ -17,6 +17,7 @@
 #define SRC_RUNTIME_LOCAL_KERNELS_MAP_EXTERNAL_CTYPESMAPKERNEL_SHAREDMEM_ADDRESS_H
 
 #include <runtime/local/datastructures/DenseMatrix.h>
+#include <runtime/local/kernels/MAP_EXTERNAL/MapKernelUtils.h>
 #include <Python.h>
 #include <memory>
 #include <util/PythonInterpreter.h>
@@ -83,7 +84,7 @@ struct CtypesMapKernelSharedMemAddress<DenseMatrix<VTRes>, DenseMatrix<VTArg>> {
         uint64_t data_address_res = reinterpret_cast<uint64_t>(res_data);
         uint64_t data_address_arg = reinterpret_cast<uint64_t>(arg_data);
 
-        std::string orig_dtype_arg = get_dtype_name();
+        std::string orig_dtype_arg = get_dtype_name<VTArg>();
         std::string orig_dtype_res = orig_dtype_arg; // Assuming VTArg and VTRes have the same data type
 
         PyObject* pArgs = Py_BuildValue("KKKKiissss",
@@ -111,26 +112,6 @@ struct CtypesMapKernelSharedMemAddress<DenseMatrix<VTRes>, DenseMatrix<VTArg>> {
         }
 
         PyGILState_Release(gstate);
-    }
-
-    static std::string get_dtype_name() {
-        if (std::is_same<VTArg, float>::value) {
-            return "float32";
-        } else if (std::is_same<VTArg, double>::value) {
-            return "float64";
-        } else if (std::is_same<VTArg, int32_t>::value) {
-            return "int32";
-        } else if (std::is_same<VTArg, int64_t>::value) {
-            return "int64";
-        } else if (std::is_same<VTArg, int8_t>::value) {
-            return "int8";
-        } else if (std::is_same<VTArg, uint64_t>::value) {
-            return "uint64";
-        } else if (std::is_same<VTArg, uint8_t>::value) {
-            return "uint8";
-        } else {
-            throw std::runtime_error("Unsupported data type!");
-        }
     }
 
     static constexpr uint32_t address_upper(uint64_t data_address) {
