+++ src/runtime/local/kernels/MAP_EXTERNAL/CtypesMapKernel_sharedMem_voidPointer.h
@@ -20,6 +20,7 @@
 #define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION
 #pragma once
 #include <runtime/local/datastructures/DenseMatrix.h>
+#include <runtime/local/kernels/MAP_EXTERNAL/MapKernelUtils.h>
 #include <Python.h>
 
 // ****************************************************************************
@@ -72,7 +73,7 @@ struct Ctypes_SharedMem_VoidPtr<DenseMatrix<VTRes>, DenseMatrix<VTArg>> {
             return;
         }
 
-        std::string dtype = get_dtype_name();
+        std::string dtype = get_dtype_name<VTArg>();
 
         PyObject* pArgs = PyTuple_Pack(7,
                                         PyLong_FromVoidPtr(arg->getValuesSharedPtr().get()),
@@ -97,26 +98,5 @@ struct Ctypes_SharedMem_VoidPtr<DenseMatrix<VTRes>, DenseMatrix<VTArg>> {
         PyGILState_Release(gstate);
 
     }
-
-    static std::string get_dtype_name() {
-        if (std::is_same<VTArg, float>::value) {
-            return "float32";
-        } else if (std::is_same<VTArg, double>::value) {
-            return "float64";
-        } else if (std::is_same<VTArg, int32_t>::value) {
-            return "int32";
-        } else if (std::is_same<VTArg, int64_t>::value) {
-            return "int64";
-        } else if (std::is_same<VTArg, int8_t>::value) {
-            return "int8";
-        } else if (std::is_same<VTArg, uint64_t>::value) {
-            return "uint64";
-        } else if (std::is_same<VTArg, uint8_t>::value) {
-            return "uint8";
-        } else {
-            throw std::runtime_error("Unsupported data type!");
-        }
-    }
 };
-
 #endif //SRC_RUNTIME_LOCAL_KERNELS_MAP_EXTERNAL_NUMPYMAPKERNEL_H
